Math operands should be cast before assignment

+++ jdwp-over-ws-client/src/main/java/com/famaridon/tcp/over/ws/client/console/StreamCountProxyListener.java
@@ -25,7 +25,7 @@ public class StreamCountProxyListener implements ProxyListener {
   }
 
   private BigInteger getByteBufferSize(ByteBuffer byteBuffer) {
-    return BigInteger.valueOf(byteBuffer.limit() - byteBuffer.position());
+    return BigInteger.valueOf((long)byteBuffer.limit() - (long)byteBuffer.position());
   }
 
   public BigInteger getReceivedFromWebSocket() {
